minor log output changes to pass the type checking on OS X (where size_t is not an unsigned int)

+++ storage/BerkeleyDBStore.cc
@@ -523,7 +523,7 @@ BerkeleyDBTable::put(const SerializableObject& key,
         return DS_ERR;
     }
 
-    log_debug("put: serializing %d byte object", sizer.size());
+    log_debug("put: serializing %u byte object", (u_int)sizer.size());
 
     { // lock
         ScopeLock lock(&scratch_mutex_);
